[REFACTOR] Adding scripting for processing the fuel-sdk repository depending on whether or not it already exists

+++ scripts/jenkins-propeller-sdk-unity-android-release-build-1.sh
@@ -15,11 +15,48 @@ fi
 ROOT_PATH=`dirname $REPOSITORY_PATH`
 PROJECT_PATH=$REPOSITORY_PATH
 FUEL_SDK_PATH="$ROOT_PATH/fuel-sdk"
+FUEL_SDK_BRANCH=master
 
-# cloning the fuel-sdk repository
-rm -rf $FUEL_SDK_PATH
-if ! git clone -b master git@github.com:Grantoo/fuel-sdk.git $FUEL_SDK_PATH ; then
-    exit 1
+CLONE_FUEL_SDK=1
+
+if [ -d $FUEL_SDK_PATH ] ; then
+    cd $FUEL_SDK_PATH
+
+    if [ -d .git ] || `git rev-parse --git-dir > /dev/null 2>&1` ; then
+        CLONE_FUEL_SDK=0
+    fi
+
+    cd - > /dev/null
+fi
+
+if [ $CLONE_FUEL_SDK == 1 ] ; then
+    # cloning the fuel-sdk repository
+    rm -rf $FUEL_SDK_PATH
+
+    if ! git clone -b $FUEL_SDK_BRANCH git@github.com:Grantoo/fuel-sdk.git $FUEL_SDK_PATH ; then
+        exit 1
+    fi
+else
+    # resetting the fuel-sdk repository
+    cd $FUEL_SDK_PATH
+
+    if ! git reset HEAD --hard ; then
+        exit 1
+    fi
+
+    if ! git clean -dfx ; then
+        exit 1
+    fi
+
+    if ! git pull ; then
+        exit 1
+    fi
+
+    if ! git checkout $FUEL_SDK_BRANCH ; then
+        exit 1
+    fi
+
+    cd - > /dev/null
 fi
 
 # removing existing files to be replaced by build artifacts to prevent issues
